Crate minimint_mint

Re-exports

pub use tiered::keys::Keys;

Modules

config

db 🔒

tiered

Data structures taking into account different amount tiers

Structs

BlindToken

Coin

A cryptographic coin consisting of a token and a threshold signature by the federated mint. In this form it can oly be validated, not spent since for that the corresponding secret spend key is required.

CoinNonce

A unique coin nonce which is also a MuSig pub key so that transactions can be signed by the spent coin’s spending keys to avoid mint frontrunning.

InvalidAmountTierError

Mint

Federated mint member mint

MintShareErrors

Represents an array of mint indexes that delivered faulty shares

PartialSigResponse

Blind signature share for a SignRequest

PartiallySignedRequest

SigResponse

Blind signature for a SignRequest

SignRequest

Request to blind sign a certain amount of coins

VerificationCache

Enums

CombineError

MintError

PeerErrorType